Squashing all changes to a single commit.
Fixes: https://github.com/axboe/fio/issues/1039
io_u: fix asymptotically high time run in accordance with blocksize of fio when using rate with timeout
stat: Adding changes to check and convert rs->agg[ddir] in KBytes

io_u:
Check if the usec is taking negative values.
However it is unlikely, But still a possibility.
If this is not taken care, usec_sleep() can take a
huge nap which is capable of not exiting the process indefinitely
although the process/thread may be sleeping