Makefile/ci: Don't pass CFLAGS when linking
There was an issue where a lld (the linker from LLVM) could fail because
a C compiler flag it didn't understand was passed to it. Fix this by no
longer passing CFLAGS to the linker.

As a result of the above, the travis "cross compiled" 32 bit Linux build
was changed to set LDFLAGS (so the linker can know about the cross
compiling).

Signed-off-by: Sitsofe Wheeler <sitsofe@yahoo.com>